requirements.txt
torch>=2.0.0
torchvision>=0.15.0
transformers>=4.35.0
opencv-python>=4.8.0
numpy>=1.24.0
pillow>=10.0.0
matplotlib>=3.7.0
tqdm>=4.66.0
accelerate>=0.24.0

# Optional - for other modules in the project
pynput>=1.7.6
SpeechRecognition>=3.10.0
edge-tts>=6.1.0
pygame>=2.5.0
bitsandbytes>=0.41.0